working on making sure it works
