+++ use-shopping-cart/src/serverUtil.js
@@ -1,8 +1,6 @@
 const validateCartItems = (inventorySrc, cartDetails) => {
   const validatedItems = []
   for (const sku in cartDetails) {
-    if (!Object.prototype.hasOwnProperty.call(cartDetails, sku)) continue
-
     const product = cartDetails[sku]
     const validatedItem = inventorySrc.find(
       (currentProduct) => currentProduct.name === product.name
